app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Dilip Sharma"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
cassandra-driver = "^3.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
